pyproject.toml
[tool.poetry]
name = "led-matrix-application"
version = "0.1.0"
description = ""
authors = ["tsomic <keller.nils99@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.taskipy.tasks]
start = "bash scripts/start.sh"
dev = "bash scripts/dev.sh"
start_windows = "pwsh scripts/start.ps1"
dev_windows = "pwsh scripts/dev.ps1"
generate_spotify_cache = "bash scripts/generate_spotify_cache.sh"

[tool.poetry.dependencies]
python = "^3.9"
starlette = "^0.37.2"
hypercorn = "^0.16.0"
pytz = "^2024.1"
pyowm = "^3.3.0"
python-dotenv = "^1.0.1"
pillow = "^10.3.0"
python-multipart = "^0.0.9"
spotipy = "^2.23.0"
setuptools = "^70.0.0"
websocket-client = "^1.8.0"
rel = "^0.4.9.19"

[tool.poetry.group.dev.dependencies]
rgbmatrixemulator = "^0.11.4"
taskipy = "^1.12.2"
watchdog = {extras = ["watchmedo"], version = "^4.0.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
